requirements.txt
nltk
tqdm
scipy
matplotlib
scikit-image
Pillow
imageio
rouge-score
